fix: DB에 유저가 동의한 정책 저장하는 로직 추가

+++ src/main/java/umc/domain/member/dto/MemberReqDTO.java
@@ -17,7 +17,7 @@ public class MemberReqDTO {
     ){}
 
     public record SignUpReq(
-            List<Integer> agreedTermsIds,
+            List<Long> agreedTermsIds,
             @NotBlank
             String name,
             @NotNull
